Reorganize some function locations
Need things to import properly after last commit
